Queue of SCell Ptrs

+++ header/cell.h
@@ -12,7 +12,7 @@ typedef struct Cell
 {
     int row;
     int col;
-    double value;
+    int value;
 
 } Cell;
 
@@ -53,7 +53,7 @@ typedef struct Cell_Formula
 
 // Function Prototypes of Cell Struct and Cell Range Struct
 
-void init_cell(Cell *cell, int row, int col, double value);
+void init_cell(Cell *cell, int row, int col, int value);
 void set_cell_value(Cell *cell, double value);
 double get_cell_value(Cell *cell);
 int get_cell_row(Cell *cell);
